Renamed extension PostType class
